experiement to see if converting to virtual functions rather than pure virtual would make a different.  It didn't. Edit time screen still craps out.

+++ demo-projects/timeandtemp/context.cpp
@@ -55,7 +55,7 @@ Context::~Context()
   delete currentScreen;
 }
 
-int backoff = 0;
+byte backoff = 0;
 
 void Context::RefreshData()
 {
